requirements.txt

Flask==2.0.1
tensorflow
opencv-python==3.4.2.17
numpy==1.21.2
Pillow==8.3.2